more fixes

+++ NanoPi Code/Software/FirmwareCommunication.c
@@ -232,7 +232,7 @@ int holdWaitCount = 0;
 const int keyRequestFrequency2 = 16000;
 const double holdSeconds = 1.0;
 
-int holdWaitTime =(int) holdSeconds*1000000.0 /keyRequestFrequency2;
+int holdWaitTime = holdSeconds*1000000.0 /keyRequestFrequency2;
 
 /**
  * Processes the output of the key presses to properly interperate when
